adds a goproxy resolver
Currently we've hardcoded the proxy, this is just a wip for now and will
be configurable from the root.

This is just a resolver for now to make sure that the code style,
errors, and the small refactors look good here.

The future direction of this would to be to attempt to use
the goproxy resolver and then failing that, fall back to
GitHub (but open to suggestions here).